+++ src/index.js
@@ -68,7 +68,7 @@ HelloWorld.prototype.eventHandlers.onLaunch = function (
   console.log('HelloWorld onLaunch');
   var speechOutput = 'Welcome to the Alexa Skills Kit, you can say hello';
   var repromptText = 'You can say hello';
-  response.ask(speechOutput, repromptText);
+  return response.ask(speechOutput, repromptText);
 };
 
 HelloWorld.prototype.eventHandlers.onSessionEnded = function (
@@ -82,24 +82,24 @@ HelloWorld.prototype.eventHandlers.onSessionEnded = function (
 HelloWorld.prototype.intentHandlers = {
   // register custom intent handlers
   HelloWorldIntent: function (intent, session, response) {
-    response.tellWithCard('Hello World!', 'Hello World', 'Hello World!');
+    return response.tellWithCard('Hello World!', 'Hello World', 'Hello World!');
   },
   'AMAZON.HelpIntent': function (intent, session, response) {
-    response.ask('You can say hello to me!', 'You can say hello to me!');
+    return response.ask('You can say hello to me!', 'You can say hello to me!');
   },
   'AMAZON.CancelIntent': function (intent, session, response) {
-    response.tell('Goodbye!');
+    return response.tell('Goodbye!');
   },
   'AMAZON.StopIntent': function (intent, session, response) {
-    response.tell('Goodbye!');
+    return response.tell('Goodbye!');
   }
 };
 
 // Create the handler that responds to the Alexa Request.
-exports.handler = function (event, context) {
+exports.handler = async function (event, context) {
   // Create an instance of the HelloWorld skill.
   var helloWorld = new HelloWorld();
-  helloWorld.execute(event, context);
+  return helloWorld.execute(event, context);
 };
 
 exports.configuredSkillId = configuredSkillId;